refactor(PhotoUploader): extract util functions
