+++ src/components/create-guild/IconSelector/components/PhotoUploader.tsx
@@ -5,8 +5,8 @@ import FormErrorMessage from "components/common/FormErrorMessage"
 import GuildLogo from "components/common/GuildLogo"
 import useDropzone, { ERROR_MESSAGES } from "hooks/useDropzone"
 import { Uploader } from "hooks/usePinata/usePinata"
-import { FileError } from "react-dropzone"
 import { useFormContext, useWatch } from "react-hook-form"
+import { getWidthAndHeightFromFile, imageDimensionsValidator } from "../utils"
 
 type Props = {
   uploader: Uploader
@@ -14,46 +14,6 @@ type Props = {
   minW?: number
   minH?: number
   onGeneratedBlobChange?: (objectURL: string) => void
-}
-
-type FileWithWidthandHeight = File & { width: number; height: number }
-
-const getWidthAndHeightFromFile = (
-  file: File
-): Promise<{ width: number; height: number }> =>
-  new Promise((resolve) => {
-    const dataURL = URL.createObjectURL(file)
-    const img = new Image()
-    img.onload = () => {
-      const { width, height } = img
-      URL.revokeObjectURL(dataURL)
-      resolve({
-        width,
-        height,
-      })
-    }
-    img.src = dataURL
-  })
-
-const imageDimensionsValidator = (
-  file: FileWithWidthandHeight,
-  minW: number,
-  minH: number
-): FileError => {
-  if (typeof file.width !== "number" && typeof file.height !== "number") return null
-
-  if (file.width < minW || file.height < minH)
-    return {
-      code: "dimension-too-small",
-      message: getDimensionErrorMessage(minW, minH),
-    }
-
-  return null
-}
-
-const getDimensionErrorMessage = (minW?: number, minH?: number): string => {
-  if (minW && minH) return `Image should be at least ${minW}x${minH}px`
-  return `Image ${minW ? "width" : "height"} should be at least ${minW || minH}px`
 }
 
 const PhotoUploader = ({
@@ -80,7 +40,7 @@ const PhotoUploader = ({
 
       for (const file of filesFromEvent) {
         filePromises.push(
-          new Promise(async (resolve) => {
+          new Promise<File>(async (resolve) => {
             if (file.type.includes("svg")) {
               resolve(file)
             } else {
@@ -97,12 +57,7 @@ const PhotoUploader = ({
       return files
     },
     validator: (file) =>
-      (minW || minH) &&
-      imageDimensionsValidator(
-        file as unknown as FileWithWidthandHeight,
-        minW,
-        minH
-      ),
+      minW || minH ? imageDimensionsValidator(file, minW ?? 0, minH ?? 0) : null,
     onDrop: (accepted) => {
       if (accepted.length > 0) {
         const generatedBlob = URL.createObjectURL(accepted[0])
